Move CellPopup -> Cell
